Bug#38151970 Upgrade the libbacktrace library to the latest [patches]
Update BACKTRACE_VERSION

Patches to upstream code:
 - Most of our changes to elf.c have been fixed upstream, apply the remaining ones.
 - Don't include backtrace-supported in state.c
 - Add our own dummy config.h

Fix Copyright year in our own files (these were skipped by the automatic script)

(cherry picked from commit df0c6a406f73b60b6a6a5788f89077fbf2fca061)
(cherry picked from commit 7158916608d95313e0868f03775134aec29fb067)

+++ extra/libbacktrace/sha793921876c981/elf.c
@@ -5757,7 +5757,7 @@ elf_uncompress_lzma_block (const unsigned char *compressed,
 	      return 0;
 	    }
 
-	  chunk_size = compressed[off] << 8;
+	  chunk_size = (size_t)compressed[off] << 8;
 	  chunk_size += compressed[off + 1];
 	  ++chunk_size;
 
@@ -5798,11 +5798,11 @@ elf_uncompress_lzma_block (const unsigned char *compressed,
 	  uncompressed_chunk_start = uncompressed_offset;
 
 	  uncompressed_chunk_size = (control & 0x1f) << 16;
-	  uncompressed_chunk_size += compressed[off] << 8;
+	  uncompressed_chunk_size += (size_t)compressed[off] << 8;
 	  uncompressed_chunk_size += compressed[off + 1];
 	  ++uncompressed_chunk_size;
 
-	  compressed_chunk_size = compressed[off + 2] << 8;
+	  compressed_chunk_size = (size_t)compressed[off + 2] << 8;
 	  compressed_chunk_size += compressed[off + 3];
 	  ++compressed_chunk_size;
